Moved datepicker in to it’s own area
