+++ src/Filters/atoms/DateTimeBlock.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled, { css } from 'styled-components';
 import {format } from 'date-fns'
 
-import Icon from '../Icons/Icon';
+import Icon from '../../Icons/Icon';
 
 const Container = styled.div`
   display: flex;
